Add documentation for BBox class, done by Clark Gaebel as part of GCI 2010. I did rework the text a little.
I would like to get some careful review for this. Nielx put out the (valid) concern that he's not confident in random students with no previous Haiku experience to write the API reference.
I'd like to know if I should put online more of these tasks, or if I should stop now because the quality is too low.


git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@39686 a95241bf-73f2-0310-859d-f6bbb57e9c96