So I'm not sure if this ever worked right or if time is just measured differently in Germany, but NetworkTime was always setting my time 6 hours ahead, no matter what time zone or whether I stored the time in local time or GMT. Upon looking closer and doing some Googling I determined that the constant Axel used for the number of seconds between 1900 and 2000 was off by 6 hours, and that combined with the odd kUTCtoGMT constant resulted in the 6 hour difference.
By just replacing the system_time_difference() function with a correct constant
for the number of seconds between 1900 and 1970 and getting rid of kUTCtoGMT it
now works for me (and at least one other person on #haiku-dev, thanks
kallisti5!) For those wondering NTP time is measured from 1900 hence the need
for this conversion.

I will now look into integrating this with the Time preflet.


git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@40103 a95241bf-73f2-0310-859d-f6bbb57e9c96