Apply the patch from Pete Goodeve from #7165 and therefore fix that ticket.
Instead of forcing a click to open rect this just restarts menu tracking if the
mouse was clicked. It seems to work great and is cleaner than my solution which
is removed in this commit as well.

This still fixes the problem of accidentially selecting menu items but also
maintains the API.

Good job Pete!


git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@40306 a95241bf-73f2-0310-859d-f6bbb57e9c96