Reimplemented how People works. The previously hard-coded attribute list is now only a fallback attribute list which is install in case the Person MIME type has no attributes at all. Otherwise the GUI is completely driven by the current Person attributes, as configured (and sorted) by FileTypes. This has been on my mental TODO list since years. Finally one can add a "Cell phone" attribute in Filetypes and actually see and edit it in People. Or a birth day attribute. If I had read the previous code correctly, it was only due to a bug that People did not remove those any attributes on every launch.
I've tested this quite a bit, but it wouldn't hurt to test it some more.
The only "regression" I am aware of is that State and Zip code are not
in one line anymore. Perhaps this feature could be reintroduced by looking
at the display width and if two attributes are short, put them in one line...


git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@40381 a95241bf-73f2-0310-859d-f6bbb57e9c96