Get the HID descriptor from the already retrieved configuration info instead of always calling the device and requesting it. If no HID descriptor is found in the configuration we fall back to retrieving it manually and if that fails we fall back to a dummy report descriptor length. In the most usual case this will save one unnecessary USB control request.
Incidentally this also fixes the absolute pointing device under VirtualBox. VBox
stalled the retrieval of the HID descriptor. That alone wasn't so problematic as
in this case we just fell back to the dummy report descriptor length which
would've been large enough. But VBox also times out the next request after the
stall, which in this case was the retrieval of the report descriptor, hence no
HIDDevice could be created.


git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@42019 a95241bf-73f2-0310-859d-f6bbb57e9c96