Fix #7672 by applying humdinger's patch then adding more:
The server name is checked while typing and if it is not valid the text is made
red. Invalid entries cannot be added to the list.

Also made the server add message be the invoking message so that the enter key
works to add a server.

After adding a valid server the text is cleared.

Valid servers names are alpha-numeric with periods, dashes and underscores.
Could need some things added but that should cover 99% of the cases.



git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@42097 a95241bf-73f2-0310-859d-f6bbb57e9c96