* Fill out the wireless network join request as detailed as possible, but don't   fail when encountering missing information (like the password). This gives the   supplicant an opportunity to ask for the required information as needed. * Remove (currently broken) WEP support from the net_server. It will be   delegated to the supplicant as well, as that one already handles all the   key/password conversion.
In the absence of a supplicant the net_server can therefore only join open
networks now. It will also only attempt that if it is sure that the network in
question is actually an open network (by means of scan results or explicit
configuration) and will otherwise delegate the join request.


git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@42774 a95241bf-73f2-0310-859d-f6bbb57e9c96